+++ packages/getbloc/lib/getbloc.dart
@@ -5,4 +5,5 @@ export 'package:get/get.dart';
 export 'src/controller.dart';
 export 'src/listener_widget.dart';
 export 'src/observer_controller.dart';
+export 'src/observer_widget.dart';
 export 'src/transform_controller.dart';