+++ packages/getbloc/lib/src/listener_widget.dart
@@ -3,13 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:getbloc/src/controller.dart';
 
-/// Signature for the `listener` function which takes the `state`
-/// and is responsible for executing in response to `state` changes.
-typedef ListenerCallback = void Function(dynamic state);
-
 /// {@template listener_widget}
-/// Takes a [Controller] and a [ListenerCallback] and invokes
-/// the [listener] in response to `state` changes in the [Controller].
+/// Takes a [Controller] and the [listener]
+/// in response to `state` changes in the [Controller].
 /// It should be used for functionality that needs to occur only in response to
 /// a `state` change such as navigation, showing a `SnackBar`, showing
 /// a `Dialog`, etc...
@@ -25,18 +21,17 @@ typedef ListenerCallback = void Function(dynamic state);
 /// )
 /// ```
 /// {@endtemplate}
-class ListenerWidget extends StatefulWidget {
+class ListenerWidget<State> extends StatefulWidget {
   /// {@macro listener_widget}
   const ListenerWidget(this.controller, this.listener, {this.child});
 
   /// The [controller] whose `state` will be listened to.
   /// Whenever the [controller]'s `state` changes, [listener] will be invoked.
-  final BaseController controller;
+  final BaseController<State> controller;
 
-  /// The [ListenerCallback] which will be called on every `state` change.
   /// This [listener] should be used for any code which needs to execute
   /// in response to a `state` change.
-  final ListenerCallback listener;
+  final void Function(State) listener;
 
   /// The widget which will be rendered as a descendant of the
   /// [ListenerWidget].