+++ packages/getbloc_test/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.1] - 2021-06-13
+
+* Update getbloc 1.1.0.
+
 ## [1.0.0] - 2021-06-09
 
 * A testing library which makes it easy to test controllers. Built to be used with the GetBloc package.